backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
spleeter==2.3.0
ffmpeg-python==0.2.0